mitosheet: pull out static type functions into non-public interface functions for clarity
