+++ mitosheet/mitosheet/public_interfaces/v1/sheet_functions/types/to_boolean_series.py
@@ -11,7 +11,7 @@ from typing import Any
 
 import numpy as np
 import pandas as pd
-from mitosheet.public_interfaces.v1.sheet_functions.types.utils import (is_bool_dtype,
+from mitosheet.is_type_utils import (is_bool_dtype,
                                                    is_datetime_dtype,
                                                    is_number_dtype,
                                                    is_string_dtype)
