+++ mitosheet/mitosheet/public_interfaces/v1/sheet_functions/types/to_datetime_series.py
@@ -9,12 +9,12 @@ For going to a datetime series.
 
 
 from typing import Any
+from mitosheet.public_interfaces.v1.sheet_functions.types.utils import get_datetime_format
 
 import numpy as np
 import pandas as pd
 
-from mitosheet.public_interfaces.v1.sheet_functions.types.utils import (get_datetime_format,
-                                                   is_bool_dtype,
+from mitosheet.is_type_utils import (is_bool_dtype,
                                                    is_datetime_dtype,
                                                    is_number_dtype,
                                                    is_string_dtype)
