+++ mitosheet/mitosheet/public_interfaces/v1/sheet_functions/types/to_float_series.py
@@ -10,10 +10,11 @@ from typing import Any
 
 import numpy as np
 import pandas as pd
-from mitosheet.public_interfaces.v1.sheet_functions.types.utils import (
-    get_billion_identifier_in_string, get_million_identifier_in_string,
+from mitosheet.is_type_utils import (
     is_bool_dtype, is_datetime_dtype, is_number_dtype, is_string_dtype)
 from typing import Any
+from mitosheet.public_interfaces.v1.sheet_functions.types.utils import get_billion_identifier_in_string, get_million_identifier_in_string
+
 
 
 def convert_string_to_float(
